testing with fake mem and redirects
