+++ repo.go
@@ -2,10 +2,8 @@ package main
 
 import (
 	"database/sql"
-	"fmt"
 	"log"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/go-sql-driver/mysql"
@@ -104,56 +102,6 @@ func (sq *purldb) createRepoDB(repo RepoObj) sql.Result {
 		log.Printf("Error creating repo: %s", err.Error())
 	}
 	return result
-}
-
-func (mr *memoryRepo) AllPurls() []Purl {
-	mr.m.RLock()
-	defer mr.m.RUnlock()
-	return mr.purls[:]
-}
-
-func (mr *memoryRepo) FindPurl(id int) Purl {
-	mr.m.RLock()
-	defer mr.m.RUnlock()
-	for _, t := range mr.purls {
-		if t.Id == id {
-			return t
-		}
-	}
-	// return empty if not found
-	return Purl{}
-}
-
-func (mr *memoryRepo) FindQuery(query string) []RepoObj {
-	mr.m.RLock()
-	defer mr.m.RUnlock()
-	var ret []RepoObj
-	for _, q := range mr.repos {
-		if strings.Contains(q.Information, query) {
-			ret = append(ret, q)
-		}
-	}
-	return ret
-}
-
-func (mr *memoryRepo) CreatePurl(t Purl) {
-	mr.m.Lock()
-	defer mr.m.Unlock()
-	mr.currentID += 1
-	t.Id = mr.currentID
-	mr.purls = append(mr.purls, t)
-}
-
-func (mr *memoryRepo) DestroyPurl(id int) error {
-	mr.m.Lock()
-	defer mr.m.Unlock()
-	for i, t := range mr.purls {
-		if t.Id == id {
-			mr.purls = append(mr.purls[:i], mr.purls[i+1:]...)
-			return nil
-		}
-	}
-	return fmt.Errorf("Could not find Purl with id of %d to delete", id)
 }
 
 // Queries the database for all purls and
